chore(test): Move the codes to runtime tests
