+++ SConstruct
@@ -639,6 +639,9 @@ if selected_platform in platform_list:
     SConscript("modules/SCsub")
     SConscript("main/SCsub")
 
+if env["tools"]:
+    SConscript("tests/runtime/SCsub")
+    
     SConscript("platform/" + selected_platform + "/SCsub")  # build selected platform
 
     # Microsoft Visual Studio Project Generation
