+++ main/SCsub
@@ -34,8 +34,5 @@ env.CommandNoCache(
 env.Depends("#main/app_icon.gen.h", "#main/app_icon.png")
 env.CommandNoCache("#main/app_icon.gen.h", "#main/app_icon.png", run_in_subprocess(main_builders.make_app_icon))
 
-if env["tools"]:
-    SConscript("tests/SCsub")
-
 lib = env.add_library("main", env.main_sources)
 env.Prepend(LIBS=[lib])
