--- main/tests/SCsub
@@ -1,9 +1,0 @@
-#!/usr/bin/python
-
-Import("env")
-
-env.tests_sources = []
-env.add_source_files(env.tests_sources, "*.cpp")
-
-lib = env.add_library("tests", env.tests_sources)
-env.Prepend(LIBS=[lib])
