+++ tests/runtime/test_string.cpp
